users_service/requirements.txt
django==4.2
djangorestframework==3.15.2
django-phonenumber-field[phonenumbers]~=8.0.0
django-redis==5.4.0
psycopg2-binary
drf-yasg